.claude/skills/record-skill/scripts/requirements.txt
pyobjc>=9.0
pyobjc-framework-Quartz>=9.0
pyobjc-framework-ApplicationServices>=9.0
